revanced-assisted-patching/requirements.txt
requests         >= 2.32.3

python-gnupg     >= 0.5.2
PyAPKDownloader  >= 0.0.2